+++ lib/core/models/editor_image.dart
@@ -4,9 +4,9 @@ import 'dart:typed_data';
 // Flutter imports:
 import 'package:flutter/material.dart';
 
-// Project imports:
 import '/core/platform/io/io_helper.dart';
 import '/shared/utils/converters.dart';
+import '/shared/utils/file_constructor_utils.dart';
 
 /// Flutter EditorImage Class Documentation
 ///
@@ -87,10 +87,11 @@ class EditorImage {
   /// must not be null.
   EditorImage({
     this.byteArray,
-    this.file,
     this.networkUrl,
     this.assetPath,
-  }) : assert(
+    dynamic file,
+  })  : file = file == null ? null : ensureFileInstance(file),
+        assert(
           byteArray != null ||
               file != null ||
               networkUrl != null ||